+++ TASON.Test/SerializeTest.cs
@@ -131,7 +131,7 @@ new Regex("[\r\n]+").Replace("""
     public void DictionaryTest()
     {
         var pairs = "[A({X:1,Y:2}),1],[A({X:2,Y:4}),2]";
-        var tason = $"Dictionary({{keyValuePairs:[{pairs}]}})";
+        var tason = $"Dictionary({{pairs:[{pairs}]}})";
 
         var s = TasonSerializer.Default.Clone();
         s.Options.UseBuiltinDictionary = true;